design: 프로필 페이지 전체 레이아웃 구현하기
